Major refactor, just started...
