+++ .env.example
@@ -6,7 +6,6 @@ SECRET_KEY=do_not_use_this_in_production
 ALLOWED_HOSTS=localhost,localhost:8000
 REDIS_URL=redis://redis:6379
 # Connected help desk
-HELP_DESK_INTERFACE=help_desk_client.interfaces.HelpDeskStubbed
 HELP_DESK_CREDS=
 HALO_SUBDOMAIN=trade
 REQUIRE_ZENDESK=True
