+++ halo/halo_manager.py
@@ -2,7 +2,7 @@ import logging
 from typing import List
 
 from halo.halo_api_client import HaloAPIClient, HaloRecordNotFoundException
-from halo.interfaces import (
+from halo.data_class import (
     HelpDeskBase,
     HelpDeskComment,
     HelpDeskException,
@@ -13,38 +13,35 @@ from halo.interfaces import (
 )
 
 
-def reverse_keys(dictionary):
-    return {value: key for key, value in dictionary.items()}
+# def reverse_keys(dictionary):
+#     return {value: key for key, value in dictionary.items()}
 
 
-STATUS_MAPPING = {
-    "new": 1,
+# Zendesk status - Halo status
+ZENDESK_TO_HALO_STATUS_MAPPING = {
+    "new": [1,4],
     "open": 2,  # in progress
     "pending": 3,  # action requried
     "on-hold": 28,
     "solved": 18,  # approved  ???
     "closed": 9,
-    "4": 4,
-    "20": 20,
-    "21": 21,
-    "22": 22,
 }
 
-REVERSE_STATUS_MAPPING = reverse_keys(STATUS_MAPPING)
+# REVERSE_STATUS_MAPPING = reverse_keys(STATUS_MAPPING)
 PRIORITY_MAPPING = {
     "incident": {"low": 4, "normal": 3, "high": 2, "urgent": 1},
 }
-REVERSE_PRIORITY_MAPPING = {key: reverse_keys(value) for key, value in PRIORITY_MAPPING.items()}
-TICKET_TYPE_MAPPING = {
-    1: TicketType.INCIDENT,
-    2: TicketType.INCIDENT,
-    3: TicketType.INCIDENT,
-    24: TicketType.INCIDENT,
-    21: TicketType.INCIDENT,
-    27: TicketType.INCIDENT,
-    28: TicketType.INCIDENT,
-    29: TicketType.INCIDENT,
-}
+# REVERSE_PRIORITY_MAPPING = {key: reverse_keys(value) for key, value in PRIORITY_MAPPING.items()}
+# TICKET_TYPE_MAPPING = {
+#     1: TicketType.INCIDENT,
+#     2: TicketType.INCIDENT,
+#     3: TicketType.INCIDENT,
+#     24: TicketType.INCIDENT,
+#     21: TicketType.INCIDENT,
+#     27: TicketType.INCIDENT,
+#     28: TicketType.INCIDENT,
+#     29: TicketType.INCIDENT,
+# }
 
 logger = logging.getLogger(__name__)
 
@@ -59,270 +56,108 @@ class HaloManager(HelpDeskBase):
             client_secret=client_secret,
         )
 
-    def get_or_create_user(self, user: HelpDeskUser = None, verb: str = "GET") -> HelpDeskUser:
-        """Get or Create a new Halo user.   /PS-IGNORE
-        :param HelpDeskUser
-                full_name: string full name for Halo user.
-                email: string email address text for the Halo user.
-        :returns: HelpDeskUser instance representing Halo user.
-        """
-        if user is not None:
-            transformed_user = self.__transform_helpdesk_user_to_halo_user(user)
-            if transformed_user.get("id") and verb == "GET":
-                halo_user = self.client.get(path=f"Users/{transformed_user['id']}")
-            elif transformed_user.get("id") and verb == "POST":
-                halo_user = self.client.post(path="Users", payload=[transformed_user])
-            else:
-                halo_user = self.client.post(path="Users", payload=[transformed_user])
-        else:
-            transformed_user = self.client.get(path="agent/me")
-            halo_user = {
-                "id": transformed_user["id"],
-                "name": transformed_user["name"],
-                "emailaddress": transformed_user["email"],
-            }
+    def get_user(self, id: str):
+        halo_user = self.client.get(path=f"Users/{id}")
+        # Need to transform into a Zendesk compatible user structure
+        zendesk_user = "TODO"
+        return zendesk_user
 
-        if halo_user is None:
-            message = f"No Halo user found for {transformed_user}"  # Error log /PS-IGNORE,
-            logger.debug(message)
-            raise HelpDeskException(message)
-        return self.__transform_halo_user_to_helpdesk_user(halo_user)
+    def create_user(self, user_details: ZendeskUserDetails = {}) -> ZendeskUser:
+        # Receive Zendesk user and create user in Halo, give back Zendesk user
+        halo_payload = "TODO create payload from incoming Zendesk user details"
+        halo_user = self.client.post(path="Users", payload=[halo_payload])
+        return halo_user
 
-    def create_ticket(self, ticket: HelpDeskTicket) -> HelpDeskTicket:
-        """Create a new Zendesk ticket in response to a new user question.
-        :param ticket: HelpDeskTicket with information to create Zendesk ticket.
-        :returns: A HelpDeskTicket instance.
-        """
-        halo_ticket = self.__transform_helpdesk_to_halo_ticket(ticket)
+    def create_ticket(self, ticket_details: ZendeskTicketDetails) -> ZendeskTicket:
+        # Create tickert 
+        halo_payload = "TODO create payload from incoming Zendesk ticket details"
+        halo_response = self.client.post(path="Tickets", payload=[halo_payload])
 
-        ticket_response = self.client.post(path="Tickets", payload=[halo_ticket])
+        zendesk_ticket_response = "TODO populate with transformed Halo response"
 
-        if ticket.comment:
-            ticket_response["comment"] = self.client.post(
+        # If we have a comment we have to add it to the ticket after creation
+        if ticket_details.comment:
+            zendesk_ticket_response["comment"] = self.client.post(
                 path="Actions",
-                payload=[
-                    self.__transform_comment_to_halo_action(ticket_response["id"], ticket.comment)
-                ],
+                payload="TODO add ticket id and outcome?",
             )
 
-        return self.__transform_object_to_helpdesk_ticket(ticket_response)
+        return zendesk_ticket_response
 
-    def get_ticket(self, ticket_id: int = None) -> HelpDeskTicket:
-        """Recover the ticket by Halo ID.
-        :param ticket_id: The Halo ID of the Ticket.
-        :returns: A HelpDeskTicket instance.
-        :raises:
-            HelpDeskTicketNotFoundException: If no ticket is found.
-        """
-        logger.debug(f"Look for Ticket by is Halo ID:<{ticket_id}>")  # /PS-IGNORE
-        try:
-            if ticket_id:
-                return [
-                    self.__transform_object_to_helpdesk_ticket(
-                        self.client.get(path=f"Tickets/{ticket_id}")
-                    )
-                ]
-            else:
-                tickets = self.client.get(path="Tickets")
-                all_tickets = []
-                for ticket in tickets["tickets"]:
-                    helpdesk_ticket = self.__transform_object_to_helpdesk_ticket(ticket)
-                    all_tickets.append(helpdesk_ticket)
-                return all_tickets
+    # def get_ticket(self, ticket_id: int = None) -> HelpDeskTicket:
+    #     """Recover the ticket by Halo ID.
+    #     :param ticket_id: The Halo ID of the Ticket.
+    #     :returns: A HelpDeskTicket instance.
+    #     :raises:
+    #         HelpDeskTicketNotFoundException: If no ticket is found.
+    #     """
+    #     logger.debug(f"Look for Ticket by is Halo ID:<{ticket_id}>")  # /PS-IGNORE
+    #     try:
+    #         if ticket_id:
+    #             return [
+    #                 self.__transform_object_to_helpdesk_ticket(
+    #                     self.client.get(path=f"Tickets/{ticket_id}")
+    #                 )
+    #             ]
+    #         else:
+    #             tickets = self.client.get(path="Tickets")
+    #             all_tickets = []
+    #             for ticket in tickets["tickets"]:
+    #                 helpdesk_ticket = self.__transform_object_to_helpdesk_ticket(ticket)
+    #                 all_tickets.append(helpdesk_ticket)
+    #             return all_tickets
 
-        except HaloRecordNotFoundException:
-            message = f"Could not find Halo ticket with ID:<{ticket_id}>"  # /PS-IGNORE
+    #     except HaloRecordNotFoundException:
+    #         message = f"Could not find Halo ticket with ID:<{ticket_id}>"  # /PS-IGNORE
 
-            logger.debug(message)
-            raise HelpDeskTicketNotFoundException(message)
+    #         logger.debug(message)
+    #         raise HelpDeskTicketNotFoundException(message)
 
-    def close_ticket(self, ticket_id: int) -> HelpDeskTicket:
-        """Close a ticket in Halo.
-        :param ticket_id: The Halo ticket ID.
-        :returns: HelpDeskTicket instance.
-        """
-        raise NotImplementedError()
+    # def close_ticket(self, ticket_id: int) -> HelpDeskTicket:
+    #     """Close a ticket in Halo.
+    #     :param ticket_id: The Halo ticket ID.
+    #     :returns: HelpDeskTicket instance.
+    #     """
+    #     raise NotImplementedError()
 
-    def add_comment(self, ticket_id: int, comment: HelpDeskComment) -> HelpDeskTicket:
-        """Add a comment to an existing ticket.
-        :param ticket_id: id of Halo ticket instance.
-        :param comment: HelpDeskComment instance.
-        :returns: The updated HelpDeskTicket instance.
-        """
-        raise NotImplementedError()
+    # def add_comment(self, ticket_id: int, comment: HelpDeskComment) -> HelpDeskTicket:
+    #     """Add a comment to an existing ticket.
+    #     :param ticket_id: id of Halo ticket instance.
+    #     :param comment: HelpDeskComment instance.
+    #     :returns: The updated HelpDeskTicket instance.
+    #     """
+    #     raise NotImplementedError()
 
-    def update_ticket(self, ticket: HelpDeskTicket) -> HelpDeskTicket:
-        """Update an existing ticket.
-        :param ticket: HelpDeskTicket ticket.
-        :returns: The updated HelpDeskTicket instance.
-        :raises:
-            HelpDeskTicketNotFoundException: If no ticket is found.
-        """
-        halo_ticket = self.__transform_helpdesk_to_halo_ticket(ticket)
-        if not ticket.id:
-            logger.error("No ticket id")
-            raise HelpDeskTicketNotFoundException("No ticket id")
+    # def update_ticket(self, ticket: HelpDeskTicket) -> HelpDeskTicket:
+    #     """Update an existing ticket.
+    #     :param ticket: HelpDeskTicket ticket.
+    #     :returns: The updated HelpDeskTicket instance.
+    #     :raises:
+    #         HelpDeskTicketNotFoundException: If no ticket is found.
+    #     """
+    #     halo_ticket = self.__transform_helpdesk_to_halo_ticket(ticket)
+    #     if not ticket.id:
+    #         logger.error("No ticket id")
+    #         raise HelpDeskTicketNotFoundException("No ticket id")
 
-        updated_ticket = self.client.post(path="Tickets", payload=[halo_ticket])
-        if updated_ticket is None:
-            message = f"Could not update ticket with id {ticket.id}"
-            logger.error(message)
-            raise HelpDeskTicketNotFoundException(message)
+    #     updated_ticket = self.client.post(path="Tickets", payload=[halo_ticket])
+    #     if updated_ticket is None:
+    #         message = f"Could not update ticket with id {ticket.id}"
+    #         logger.error(message)
+    #         raise HelpDeskTicketNotFoundException(message)
 
-        if halo_ticket["comment"]:
-            updated_ticket["comment"] = self.client.post(
-                path="Actions", payload=[halo_ticket["comment"]]
-            )
+    #     if halo_ticket["comment"]:
+    #         updated_ticket["comment"] = self.client.post(
+    #             path="Actions", payload=[halo_ticket["comment"]]
+    #         )
 
-        return self.__transform_object_to_helpdesk_ticket(updated_ticket)
+    #     return self.__transform_object_to_helpdesk_ticket(updated_ticket)
 
-    def get_comments(self, ticket_id: int) -> List[HelpDeskComment]:
-        comments = []
-        ticket_actions = self.client.get(f"Actions?ticket_id={ticket_id}")
-        for action in reversed(ticket_actions["actions"]):
-            if action["outcome"] == "comment":
-                comment = self.__transform_halo_action_to_comment(action)
-                comments.append(comment)
-        return comments
-
-    def __transform_comment_to_halo_action(self, ticket_id: int, comment: HelpDeskComment):
-        """
-        :param ticket_id: id of the ticket comment
-        :param comment: HelpDeskComment instance.
-        :returns: halo action object.
-        """
-        return {
-            "note": comment[0]["body"],
-            "ticket_id": ticket_id,
-            "hiddenfromuser": not comment[0]["public"],
-            # "who_agentid": comment.author_id,
-            "outcome": "comment",
-        }
-
-    def __transform_halo_action_to_comment(self, halo_comment) -> HelpDeskComment:
-        """Transform from halo comment format to HelpDeskComment.
-        :param halo_comment: Halo Action instance.
-        :returns: halo action object.
-        """
-        return HelpDeskComment(
-            body=halo_comment["note"],
-            public=not halo_comment["hiddenfromuser"],
-            author_id=halo_comment.get("who_agentid"),
-        )
-
-    def __transform_helpdesk_to_halo_ticket(self, ticket: HelpDeskTicket) -> object:
-        """
-        :param ticket: HelpDeskTicket instance.
-        :returns: halo ticket object.
-        """
-        # custom_fields = None
-
-        ticket_user = self.get_or_create_user(ticket.user)
-
-        # if ticket.custom_fields:
-        #     custom_fields = [
-        #         {'id':custom_field.id, 'value':custom_field.value}
-        #         for custom_field in ticket.custom_fields
-        #     ]
-
-        return {
-            "id": ticket.id,
-            "status_id": STATUS_MAPPING[ticket.status] if ticket.status else None,
-            "priority_id": PRIORITY_MAPPING[ticket.ticket_type.value][ticket.priority]
-            if (ticket.ticket_type and ticket.priority)
-            else None,
-            # "emailcclist": [ticket.recipient_email], #correct?
-            "summary": ticket.subject,
-            "details": ticket.description,
-            "user_id": ticket_user.id,
-            "agent_id": ticket.assignee_id,
-            "team_id": ticket.group_id,
-            "ticket_tags": ", ".join(ticket.tags) if ticket.tags else None,
-            # ticket.service_name field be determined
-            "third_party_id": ticket.external_id,  # /PS-IGNORE
-            # TODO 'custom_fields':custom_fields, #+ [{'id':144,'name':
-            # 'CFemailAddress', 'value': ticket.recipient_email }],
-            "tickettype_id": TICKET_TYPE_MAPPING[ticket.ticket_type]
-            if ticket.ticket_type
-            else None,
-            "comment": self.__transform_comment_to_halo_action(ticket.id, ticket.comment)
-            if ticket.id and ticket.comment
-            else None,
-            "category_1": "Standard Applications>Adobe",
-        }
-
-    def __transform_object_to_helpdesk_ticket(self, ticket: object) -> HelpDeskTicket:
-        """Transform Halo ticket object into HelpDeskTicket instance.
-        :param ticket: Halo ticket object.
-        :returns: HelpDeskTicket instance.
-        """
-        ticket_user, comment = None, None
-
-        if ticket.get("user"):
-            ticket_user = HelpDeskUser(
-                id=ticket["user"]["id"],
-                full_name=ticket["user"]["name"],
-                email=ticket["user"]["name"],
-            )
-        elif ticket.get("user_id"):
-            ticket_user = HelpDeskUser(id=ticket["user_id"])
-
-        # Get ticket comments if comments set latest comment
-        ticket_actions = self.client.get(f"Actions?ticket_id={ticket['id']}")
-        comments = []
-        for action in reversed(ticket_actions["actions"]):
-            if action["outcome"] == "comment":
-                comment = self.__transform_halo_action_to_comment(action)
-                comments.append(comment)
-
-        helpdesk_ticket = HelpDeskTicket(
-            id=ticket["id"],
-            status=REVERSE_STATUS_MAPPING[ticket["status_id"]] if ticket.get("status_id") else None,
-            subject=ticket["summary"] if "summary" in ticket else "",
-            description=ticket.get("details", None),
-            # recipient_email=getattr(ticket, "recipient", None),
-            user=ticket_user,
-            created_at=ticket.get("dateoccurred"),
-            updated_at=ticket.get("flastupdate"),
-            priority=REVERSE_PRIORITY_MAPPING[TICKET_TYPE_MAPPING[ticket["tickettype_id"]].value][
-                ticket.get("priority_id")
-            ],
-            due_at=ticket.get("deadlinedate", None),
-            assignee_id=ticket.get("agent_id", None),
-            group_id=ticket.get("team_id", None),
-            external_id=ticket.get("third_party_id", None),  # /PS-IGNORE
-            tags=(ticket.get("ticket_tags", [])).split(", ") if "ticket_tags" in ticket else [],
-            # custom_fields=custom_fields,
-            comment=comments,
-            ticket_type=TICKET_TYPE_MAPPING[ticket["tickettype_id"]]
-            if ticket.get("tickettype_id")
-            else None,
-        )
-        return helpdesk_ticket
-
-    def __transform_helpdesk_user_to_halo_user(self, user: HelpDeskUser) -> object:
-        """Transform HelpDesk user into Halo user object.
-        :param user: HelpDeskUser instance.
-        :returns: Halo User object.
-        """
-        if user.id:
-            return {"id": user.id, "name": user.full_name, "emailaddress": user.email}
-        elif user.id is None:
-            return {"name": user.full_name, "emailaddress": user.email, "site_id": user.site_id}
-        else:
-            # This should not be possible so raise exception
-            raise HelpDeskException(
-                "Cannot transform user to Halo user",
-            )
-
-    def __transform_halo_user_to_helpdesk_user(self, user: object) -> HelpDeskUser:
-        """Transform Halo user object into HelpDesk user instance.
-        :param user: Halo user object
-        :returns: HelpDeskUser user instance.
-        """
-        localuser = HelpDeskUser(id=user["id"], full_name=user["name"])
-        if "emailaddress" in user:
-            localuser.email = user["emailaddress"]
-            return localuser
-        else:
-            return localuser
+    # def get_comments(self, ticket_id: int) -> List[HelpDeskComment]:
+    #     comments = []
+    #     ticket_actions = self.client.get(f"Actions?ticket_id={ticket_id}")
+    #     for action in reversed(ticket_actions["actions"]):
+    #         if action["outcome"] == "comment":
+    #             comment = self.__transform_halo_action_to_comment(action)
+    #             comments.append(comment)
+    #     return comments
