+++ help_desk_api/views.py
@@ -1,5 +1,5 @@
 from halo.halo_manager import HaloManager
-from halo.interfaces import HelpDeskTicket, HelpDeskUser
+from halo.data_class import HelpDeskTicket, HelpDeskUser
 from rest_framework import authentication, permissions, status
 from rest_framework.renderers import BrowsableAPIRenderer, JSONRenderer
 from rest_framework.response import Response
@@ -37,15 +37,17 @@ class UserView(HaloBaseView):
 
     def get(self, request, *args, **kwargs):
         """
-        GET User in Halo
+        Get a user from Halo
         """
         # Build User based on ID
 
         if self.kwargs.get("id"):
-            user = HelpDeskUser(id=self.kwargs.get("id"))
             # Get User from Halo
-            queryset = self.halo_manager.get_or_create_user(user, request.method)
-            serializer = HaloUserSerializer(queryset)
+            halo_user = self.halo_manager.get_user(id=self.kwargs.get("id"))
+            serializer = ZendeskUserSerializer(
+                id=halo_user["id"],
+                ...
+            )
             return Response(serializer.data)
         else:
             # if no user id is passed we show the agent me??
@@ -53,103 +55,98 @@ class UserView(HaloBaseView):
 
     def post(self, request, *args, **kwargs):
         """
-        CREATE/UPDATE a User in Halo
+        Create a User in Halo
         """
-        # Create/Update User in Halo
-        user = HelpDeskUser(
-            full_name=request.data["name"],
-            email=request.data["emailaddress"],
-        )
-        if "id" in request.data:
-            user.id = request.data["id"]
-            queryset = self.halo_manager.get_or_create_user(user, request.method)
-        else:
-            user.site_id = request.data["site_id"]
-            queryset = self.halo_manager.get_or_create_user(user, request.method)
-        serializer = HaloUserSerializer(queryset)
+        # Create user in Halo
+        halo_user = self.halo_manager.create_user(request.data) # ** maybe?
+        serializer = ZendeskUserSerializer(halo_user)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 
-class MeView(HaloBaseView):
-    """
-    View for interaction with self
-    """
+# class MeView(HaloBaseView):
+#     """
+#     View for interaction with self
+#     """
 
-    authentication_classes = [authentication.TokenAuthentication]
-    permission_classes = [permissions.AllowAny]
-    renderer_classes = [BrowsableAPIRenderer, JSONRenderer]
+#     authentication_classes = [authentication.TokenAuthentication]
+#     permission_classes = [permissions.AllowAny]
+#     renderer_classes = [BrowsableAPIRenderer, JSONRenderer]
 
-    def get(self, request, format=None):
-        """
-        GET Agent Me in Halo
-        """
-        # Get Me from Halo
-        queryset = self.halo_manager.get_or_create_user(user=None)
-        serializer = HaloUserSerializer(queryset)
-        return Response(serializer.data)
-
-
-class CommentView(HaloBaseView):
-    """
-    View for interaction with comment
-    """
-
-    authentication_classes = [authentication.TokenAuthentication]
-    permission_classes = [permissions.AllowAny]
-    renderer_classes = [BrowsableAPIRenderer, JSONRenderer]
-
-    def get(self, request, id, format=None):
-        """
-        GET comments from Halo
-        """
-        # Get ticket from Halo
-        queryset = self.halo_manager.get_comments(ticket_id=id)
-        serializer = CommentSerializer(queryset, many=True)
-        return Response(serializer.data)
+#     def get(self, request, format=None):
+#         """
+#         GET Agent Me in Halo
+#         """
+#         # Get Me from Halo
+#         queryset = self.halo_manager.get_or_create_user(user=None)
+#         serializer = HaloUserSerializer(queryset)
+#         return Response(serializer.data)
 
 
-class TicketView(HaloBaseView):
-    """
-    View for interacting with tickets
-    """
+# class CommentView(HaloBaseView):
+#     """
+#     View for interaction with comment
+#     """
 
-    authentication_classes = [authentication.TokenAuthentication]
-    permission_classes = [permissions.AllowAny]
-    serializer_class = TicketSerializer
-    renderer_classes = [BrowsableAPIRenderer, JSONRenderer]
+#     authentication_classes = [authentication.TokenAuthentication]
+#     permission_classes = [permissions.AllowAny]
+#     renderer_classes = [BrowsableAPIRenderer, JSONRenderer]
 
-    def get(self, request, *args, **kwargs):
-        """
-        GET ticket/tickets from Halo
-        """
-        if self.kwargs.get("id"):
-            queryset = self.halo_manager.get_ticket(ticket_id=self.kwargs.get("id"))
-        else:
-            queryset = self.halo_manager.get_ticket()
-        serializer = TicketSerializer(queryset, many=True)
-        return Response(serializer.data)
+#     def get(self, request, id, format=None):
+#         """
+#         GET comments from Halo
+#         """
+#         # Get ticket from Halo
+#         queryset = self.halo_manager.get_comments(ticket_id=id)
+#         serializer = CommentSerializer(queryset, many=True)
+#         return Response(serializer.data)
 
-    def post(self, request, *args, **kwargs):
-        """
-        CREATE/UPDATE ticket in Halo
-        """
-        ticket = HelpDeskTicket(
-            subject=request.data["subject"],
-            description=request.data["description"],
-            priority=request.data["priority"],
-            comment=request.data["comment"],
-            user=None,
-        )
 
-        if "id" in request.data:
-            # update the ticket
-            ticket.id = request.data["id"]
-            queryset = self.halo_manager.update_ticket(
-                ticket,
-            )
-        else:
-            queryset = self.halo_manager.create_ticket(
-                ticket,
-            )
-        serializer = TicketSerializer(queryset)
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
+# class TicketView(HaloBaseView):
+#     """
+#     View for interacting with tickets
+#     """
+
+#     authentication_classes = [authentication.TokenAuthentication]
+#     permission_classes = [permissions.AllowAny]
+#     serializer_class = TicketSerializer
+#     renderer_classes = [BrowsableAPIRenderer, JSONRenderer]
+
+#     def get(self, request, *args, **kwargs):
+#         """
+#         GET ticket/tickets from Halo
+#         """
+#         if self.kwargs.get("id"):
+#             queryset = self.halo_manager.get_ticket(ticket_id=self.kwargs.get("id"))
+#         else:
+#             queryset = self.halo_manager.get_ticket()
+#         serializer = TicketSerializer(queryset, many=True)
+#         return Response(serializer.data)
+
+#     def post(self, request, *args, **kwargs):
+#         """
+#         CREATE/UPDATE ticket in Halo
+#         """
+#         ticket = HelpDeskTicket(
+#             subject=request.data["subject"],
+#             description=request.data["description"],
+#             priority=request.data["priority"],
+#             comment=request.data["comment"],
+#             user=None,
+#         )
+
+#         if "id" in request.data:
+#             # update the ticket
+#             ticket.id = request.data["id"]
+
+
+
+
+#             queryset = self.halo_manager.update_ticket(
+#                 ticket,
+#             )
+#         else:
+#             queryset = self.halo_manager.create_ticket(
+#                 ticket,
+#             )
+#         serializer = TicketSerializer(queryset)
+#         return Response(serializer.data, status=status.HTTP_201_CREATED)
